+++ packages/assignment-7/src/dateUtils.ts
@@ -1,5 +1,15 @@
 import { Event } from "./type/schedule.type";
 
+export const weekDays = ["일", "월", "화", "수", "목", "금", "토"];
+
+export const notificationOptions = [
+  { value: 1, label: "1분 전" },
+  { value: 10, label: "10분 전" },
+  { value: 60, label: "1시간 전" },
+  { value: 120, label: "2시간 전" },
+  { value: 1440, label: "1일 전" },
+];
+
 export const getDaysInMonth = (year: number, month: number) => {
   return new Date(year, month + 1, 0).getDate();
 };
